move APIFnGetBridges to action_getbridges.go
